+++ app/ThirdPartySupport/Scf.php
@@ -72,12 +72,12 @@ class Scf extends Base implements ThirdPartySupport_Base {
 	 * @return void
 	 */
 	public function init(): void {
-		// bail if plugin is not enabled.
+		// bail if the plugin is not enabled.
 		if ( ! $this->is_active() ) {
 			return;
 		}
 
-		// use ACF hooks.
+		// use SCF hooks.
 		add_action( 'acf/render_field_settings', array( $this, 'extend_field_settings' ) );
 
 		// use our own hooks.
@@ -148,7 +148,12 @@ class Scf extends Base implements ThirdPartySupport_Base {
 	 * @return void
 	 */
 	public function add_meta_fields_to_simplification( Post_Object $post_object, int $post_id, string $source_language ): void {
-		// get all ACF fields on this object.
+		// bail if SCF functions are not available.
+		if ( ! function_exists( 'get_fields' ) ) {
+			return;
+		}
+
+		// get all SCF fields on this object.
 		$fields = get_fields( $post_object->get_id() );
 
 		// bail if no fields are configured.
@@ -213,12 +218,12 @@ class Scf extends Base implements ThirdPartySupport_Base {
 		);
 
 		/**
-		 * Filter the list of simplifiable fields in ACF.
+		 * Filter the list of simplifiable fields in SCF.
 		 *
 		 * @since 3.2.0 Available since 3.2.0.
 		 * @param array<string,bool> $types List of types.
 		 */
-		return apply_filters( 'easy_language_acf_types', $types );
+		return apply_filters( 'easy_language_scf_types', $types );
 	}
 
 	/**
@@ -253,8 +258,8 @@ class Scf extends Base implements ThirdPartySupport_Base {
 	 * @return void
 	 */
 	public function extend_field_settings( array $field ): void {
-		// bail if this field is not supported.
-		if ( ! array_key_exists( $field['type'], $this->get_types() ) ) {
+		// bail if this field is not supported or main SCF-function is not available.
+		if ( ! array_key_exists( $field['type'], $this->get_types() ) || ! function_exists( 'acf_render_field_setting' ) ) {
 			return;
 		}